+++ commands/moderation/ban.js
@@ -2,17 +2,17 @@ const {RichEmbed}=require("discord.js")
 
 module.exports={
 	run:async(client,database,config,message,args)=>{
-		if(!message.member.hasPermission(['BAN_MEMBERS','ADMINISTRATOR']))
+		if(!message.member.hasPermission(['ADMINISTRATOR']))
 			return message.channel.send('```Nie możesz użyć tej komendy.```')
 		const member=message.mentions.members.first()
 		if(!member)
 			return message.channel.send('```Nie znaleziono użytkownika do zbanowania.\nUżycie: !ban użytkownik powód```')
 		const reason=args.slice(1).join(' ')||'Brak'
-		member.send(`Cześć, zostałeś zbanowany na **${message.guild.name}**.\nPowód: ${reason}.`)
+		await member.send(`Cześć, zostałeś zbanowany na **${message.guild.name}**.\nPrzez: **${message.author.tag}**\nPowód: **${reason}**.`).catch(()=>{})
 		member.ban(reason).then(()=>{
 			message.channel.send(new RichEmbed()
 				.setColor(0xFFFF00)
-				.setTitle(`Użytkownik ${member.user.tag} został zbanowany.`)
+				.setTitle(`Użytkownik **${member.user.tag}** został zbanowany.`)
 				.addField('Przez:',message.author.tag)
 				.addField('Powód:',reason)
 			)