+++ commands/moderation/unban.js
@@ -2,7 +2,7 @@ const {RichEmbed}=require("discord.js")
 
 module.exports={
 	run:async(client,database,config,message,args)=>{
-		if(!message.member.hasPermission(['BAN_MEMBERS','ADMINISTRATOR']))
+		if(!message.member.hasPermission(['ADMINISTRATOR']))
 			return message.channel.send('```Nie możesz użyć tej komendy.```')
 		if(!args[0])
 			return message.channel.send('```Nie znaleziono użytkownika do odbanowania.\nUżycie: !unban użytkownik```')
@@ -10,7 +10,7 @@ module.exports={
 		message.guild.unban(user).then(()=>{
 			message.channel.send(new RichEmbed()
 				.setColor(0x00FF00)
-				.setTitle(`Użytkownik ${user.tag} został odbanowany.`)
+				.setTitle(`Użytkownik **${user.tag}** został odbanowany.`)
 				.addField('Przez:',message.author.tag)
 			)
 		}).catch((err)=>{