+++ index.js
@@ -5,7 +5,7 @@ const {basename}=require('path')
 const {Client}=require('discord.js')
 const client=new Client()
 const config=require('./config.json')
-const database=require('./database.json')
+let database=require('./database.json')
 let commands=[]
 glob('commands/**/*.js',(err,files)=>{
 	if(err)console.log(err)
@@ -39,6 +39,17 @@ client.on('message',async message=>{
 		message.delete()
 		if(command in commands){
 			commands[command].run(client,database,config,message,args)
+		}else{
+			if(!message.member.id==config.ownerID)
+				message.channel.send('```Nie możesz użyć tej komendy.```')
+			switch(command){
+				case 'dbsave':
+					writeFile('database.json',JSON.stringify(database,null,4),function(){})
+					break;
+				case 'dbload':
+					database=require('./database.json')
+					break;
+			}
 		}
 	}
 })